!!! TASK: Refactor node label rendering (as pure Neos' concept)
The CR core will not know anything about node labels anymore.
This will be a pure Neos' attached concept.
