+++ Neos.ContentRepositoryRegistry/Classes/Factory/NodeTypeManager/DefaultNodeTypeManagerFactory.php
@@ -11,7 +11,6 @@ readonly class DefaultNodeTypeManagerFactory implements NodeTypeManagerFactoryIn
 {
     public function __construct(
         private ConfigurationManager $configurationManager,
-        private ObjectManagerBasedNodeLabelGeneratorFactory $nodeLabelGeneratorFactory,
         private NodeTypeEnrichmentService $nodeTypeEnrichmentService,
     ) {
     }
@@ -23,8 +22,7 @@ readonly class DefaultNodeTypeManagerFactory implements NodeTypeManagerFactoryIn
             function () {
                 $configuration = $this->configurationManager->getConfiguration('NodeTypes');
                 return $this->nodeTypeEnrichmentService->enrichNodeTypeLabelsConfiguration($configuration);
-            },
-            $this->nodeLabelGeneratorFactory
+            }
         );
     }
 }
