+++ Neos.Neos/Classes/Domain/NodeLabel/ExpressionBasedNodeLabelGenerator.php
@@ -1,5 +1,5 @@
 <?php
-namespace Neos\ContentRepositoryRegistry\NodeLabel;
+namespace Neos\Neos\Domain\NodeLabel;
 
 /*
  * This file is part of the Neos.ContentRepository package.
@@ -11,7 +11,6 @@ namespace Neos\ContentRepositoryRegistry\NodeLabel;
  * source code.
  */
 
-use Neos\ContentRepository\Core\NodeType\NodeLabelGeneratorInterface;
 use Neos\ContentRepository\Core\Projection\ContentGraph\Node;
 use Neos\Eel\EelEvaluatorInterface;
 use Neos\Eel\Utility;
