+++ Neos.Neos/Tests/Functional/Service/NodeTypeSchemaBuilderTest.php
@@ -11,7 +11,6 @@ namespace Neos\Neos\Tests\Functional\Service;
  * source code.
  */
 
-use Neos\ContentRepository\Core\NodeType\NodeLabelGeneratorFactoryInterface;
 use Neos\ContentRepository\Core\NodeType\NodeTypeManager;
 use Neos\Flow\Configuration\ConfigurationManager;
 use Neos\Flow\Tests\FunctionalTestCase;
@@ -38,11 +37,9 @@ class NodeTypeSchemaBuilderTest extends FunctionalTestCase
     {
         parent::setUp();
         $configurationManager = $this->objectManager->get(ConfigurationManager::class);
-        $nodeLabelGeneratorFactory = $this->objectManager->get(NodeLabelGeneratorFactoryInterface::class);
         $this->nodeTypeSchemaBuilder = NodeTypeSchemaBuilder::create(
             new NodeTypeManager(
-                fn() => $configurationManager->getConfiguration('NodeTypes'),
-                $nodeLabelGeneratorFactory
+                fn() => $configurationManager->getConfiguration('NodeTypes')
             )
         );
         $this->schema = $this->nodeTypeSchemaBuilder->generateNodeTypeSchema();
